Dönüşüm iyileştirmeleri yapıldı

+++ KontrolPaneli/src/Model/Motor/KalipMotor.java
@@ -8,7 +8,7 @@ public class KalipMotor extends Motor {
 		super();
 		this.kasnakOran = MachineConfig.kalipKasnakOran;
 		this.stepCountPerRev = MachineConfig.stepPerRev;
-		this.oneMMStepcount = (long)((this.kasnakOran * this.stepCountPerRev) / MachineConfig.milHatve); // cast kısmında hata çıkabilir buraya dikkat et
+		this.oneMMStepcount = ((this.kasnakOran * this.stepCountPerRev) / MachineConfig.milHatve); // cast kısmında hata çıkabilir buraya dikkat et
 		
 	}
